example/resource-provider-api/requirements.txt
django==2.0
git+git://github.com/Peter-Slump/django-keycloak.git#egg=django-keycloak
django-dynamic-fixtures==0.1.7